+++ benchmarks/PerformanceBenchmark.php
@@ -22,6 +22,10 @@ class PerformanceBenchmark
         '16KB' => 16 * 1024,
         '32KB' => 32 * 1024,
         '64KB' => 64 * 1024,
+        '128KB' => 128 * 1024,
+        '256KB' => 256 * 1024,
+        '512KB' => 512 * 1024,
+        '1MB' => 1024 * 1024,
     ];
 
     public function __construct()